// give-away/src/api/models/FashionItemDetailResponseOrderDetailResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Give Away API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { FashionItemDetailResponse } from './FashionItemDetailResponse';
import {
    FashionItemDetailResponseFromJSON,
    FashionItemDetailResponseFromJSONTyped,
    FashionItemDetailResponseToJSON,
} from './FashionItemDetailResponse';

/**
 * 
 * @export
 * @interface FashionItemDetailResponseOrderDetailResponse
 */
export interface FashionItemDetailResponseOrderDetailResponse {
    /**
     * 
     * @type {string}
     * @memberof FashionItemDetailResponseOrderDetailResponse
     */
    orderDetailId?: string;
    /**
     * 
     * @type {number}
     * @memberof FashionItemDetailResponseOrderDetailResponse
     */
    unitPrice?: number;
    /**
     * 
     * @type {string}
     * @memberof FashionItemDetailResponseOrderDetailResponse
     */
    orderId?: string;
    /**
     * 
     * @type {Date}
     * @memberof FashionItemDetailResponseOrderDetailResponse
     */
    refundExpirationDate?: Date | null;
    /**
     * 
     * @type {FashionItemDetailResponse}
     * @memberof FashionItemDetailResponseOrderDetailResponse
     */
    fashionItemDetail?: FashionItemDetailResponse;
}

/**
 * Check if a given object implements the FashionItemDetailResponseOrderDetailResponse interface.
 */
export function instanceOfFashionItemDetailResponseOrderDetailResponse(value: object): value is FashionItemDetailResponseOrderDetailResponse {
    return true;
}

export function FashionItemDetailResponseOrderDetailResponseFromJSON(json: any): FashionItemDetailResponseOrderDetailResponse {
    return FashionItemDetailResponseOrderDetailResponseFromJSONTyped(json, false);
}

export function FashionItemDetailResponseOrderDetailResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): FashionItemDetailResponseOrderDetailResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'orderDetailId': json['orderDetailId'] == null ? undefined : json['orderDetailId'],
        'unitPrice': json['unitPrice'] == null ? undefined : json['unitPrice'],
        'orderId': json['orderId'] == null ? undefined : json['orderId'],
        'refundExpirationDate': json['refundExpirationDate'] == null ? undefined : (new Date(json['refundExpirationDate'])),
        'fashionItemDetail': json['fashionItemDetail'] == null ? undefined : FashionItemDetailResponseFromJSON(json['fashionItemDetail']),
    };
}

export function FashionItemDetailResponseOrderDetailResponseToJSON(value?: FashionItemDetailResponseOrderDetailResponse | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'orderDetailId': value['orderDetailId'],
        'unitPrice': value['unitPrice'],
        'orderId': value['orderId'],
        'refundExpirationDate': value['refundExpirationDate'] == null ? undefined : ((value['refundExpirationDate'] as any).toISOString()),
        'fashionItemDetail': FashionItemDetailResponseToJSON(value['fashionItemDetail']),
    };
}