// give-away/src/api/models/FashionItemOrderDetailResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Give Away API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { FashionItem } from './FashionItem';
import {
    FashionItemFromJSON,
    FashionItemFromJSONTyped,
    FashionItemToJSON,
} from './FashionItem';

/**
 * 
 * @export
 * @interface FashionItemOrderDetailResponse
 */
export interface FashionItemOrderDetailResponse {
    /**
     * 
     * @type {string}
     * @memberof FashionItemOrderDetailResponse
     */
    orderDetailId?: string;
    /**
     * 
     * @type {number}
     * @memberof FashionItemOrderDetailResponse
     */
    unitPrice?: number;
    /**
     * 
     * @type {string}
     * @memberof FashionItemOrderDetailResponse
     */
    orderId?: string;
    /**
     * 
     * @type {Date}
     * @memberof FashionItemOrderDetailResponse
     */
    refundExpirationDate?: Date | null;
    /**
     * 
     * @type {FashionItem}
     * @memberof FashionItemOrderDetailResponse
     */
    fashionItemDetail?: FashionItem;
}

/**
 * Check if a given object implements the FashionItemOrderDetailResponse interface.
 */
export function instanceOfFashionItemOrderDetailResponse(value: object): value is FashionItemOrderDetailResponse {
    return true;
}

export function FashionItemOrderDetailResponseFromJSON(json: any): FashionItemOrderDetailResponse {
    return FashionItemOrderDetailResponseFromJSONTyped(json, false);
}

export function FashionItemOrderDetailResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): FashionItemOrderDetailResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'orderDetailId': json['orderDetailId'] == null ? undefined : json['orderDetailId'],
        'unitPrice': json['unitPrice'] == null ? undefined : json['unitPrice'],
        'orderId': json['orderId'] == null ? undefined : json['orderId'],
        'refundExpirationDate': json['refundExpirationDate'] == null ? undefined : (new Date(json['refundExpirationDate'])),
        'fashionItemDetail': json['fashionItemDetail'] == null ? undefined : FashionItemFromJSON(json['fashionItemDetail']),
    };
}

export function FashionItemOrderDetailResponseToJSON(value?: FashionItemOrderDetailResponse | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'orderDetailId': value['orderDetailId'],
        'unitPrice': value['unitPrice'],
        'orderId': value['orderId'],
        'refundExpirationDate': value['refundExpirationDate'] == null ? undefined : ((value['refundExpirationDate'] as any).toISOString()),
        'fashionItemDetail': FashionItemToJSON(value['fashionItemDetail']),
    };
}